+++ app/build.gradle
@@ -3,6 +3,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'dagger.hilt.android.plugin'
 
 android {
     compileSdkVersion versions.compileSdk
@@ -83,12 +84,15 @@ dependencies {
     //paging Library
     implementation 'androidx.paging:paging-runtime:' + versions.paging
 
-    //dagger
-    implementation 'com.google.dagger:dagger:' + versions.dagger
-    implementation 'com.google.dagger:dagger-android:' + versions.dagger
-    implementation 'com.google.dagger:dagger-android-support:' + versions.dagger
-    kapt 'com.google.dagger:dagger-android-processor:' + versions.dagger
-    kapt 'com.google.dagger:dagger-compiler:' + versions.dagger
+    // hilt
+    implementation 'com.google.dagger:hilt-android:' + versions.daggerHiltAndroidVersion
+    implementation 'com.google.dagger:hilt-android-testing:' + versions.daggerHiltAndroidVersion
+    implementation 'androidx.hilt:hilt-common:' + versions.daggerHiltVersion
+    implementation 'androidx.hilt:hilt-lifecycle-viewmodel:' + versions.daggerHiltVersion
+    kapt 'com.google.dagger:hilt-android-compiler:' + versions.daggerHiltAndroidVersion
+    kapt 'androidx.hilt:hilt-compiler:' + versions.daggerHiltVersion
+    androidTestImplementation 'com.google.dagger:hilt-android-testing:' + versions.daggerHiltAndroidVersion
+    kaptAndroidTest 'com.google.dagger:hilt-android-compiler:' +versions.daggerHiltAndroidVersion
 
     //glide
     implementation 'com.github.bumptech.glide:glide:' + versions.glide
@@ -112,6 +116,8 @@ dependencies {
     //rx
     implementation 'io.reactivex.rxjava2:rxjava:' + versions.rxjava
     implementation 'io.reactivex.rxjava2:rxandroid:' + versions.rxandroid
+
+
 
     //navigation
     implementation 'androidx.navigation:navigation-fragment-ktx:' + versions.navigation
