--- app/src/main/java/com/xplorer/projectx/di/AppComponent.kt
@@ -1,44 +1,0 @@
-/**
- *  Designed and developed by ProjectX
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-package com.xplorer.projectx.di
-
-import android.app.Application
-import com.xplorer.projectx.ProjectX
-import com.xplorer.projectx.di.modules.AppModule
-import com.xplorer.projectx.di.modules.MainActivityModule
-import com.xplorer.projectx.di.modules.ViewModelModule
-import dagger.BindsInstance
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-import dagger.android.AndroidInjector
-import javax.inject.Singleton
-
-@Singleton
-@Component(
-  modules = [AndroidInjectionModule::class,
-    AppModule::class,
-    MainActivityModule::class,
-    ViewModelModule::class]
-)
-interface AppComponent : AndroidInjector<ProjectX> {
-  @Component.Builder
-  interface Builder {
-    @BindsInstance
-    fun application(application: Application): Builder
-
-    fun build(): AppComponent
-  }
-}
