--- app/src/main/java/com/xplorer/projectx/di/modules/MainActivityModule.kt
@@ -1,26 +1,0 @@
-/**
- *  Designed and developed by ProjectX
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-package com.xplorer.projectx.di.modules
-
-import com.xplorer.projectx.ui.MainActivity
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class MainActivityModule {
-    @ContributesAndroidInjector(modules = [SearchStartFragmentModule::class, CityFragmentModule::class])
-    internal abstract fun contributeMainActivity(): MainActivity
-}
