--- app/src/main/java/com/xplorer/projectx/di/modules/ViewModelModule.kt
@@ -1,54 +1,0 @@
-/**
- *  Designed and developed by ProjectX
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-package com.xplorer.projectx.di.modules
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.xplorer.projectx.ui.city.CityMapViewModel
-import com.xplorer.projectx.ui.city.CityPhotoViewModel
-import com.xplorer.projectx.ui.city.CitySearchViewModel
-import com.xplorer.projectx.ui.search.SearchStartViewModel
-import com.xplorer.projectx.view_model.ViewModelFactory
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class ViewModelModule {
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(CitySearchViewModel::class)
-    internal abstract fun bindSearchViewModel(citySearchViewModel: CitySearchViewModel): ViewModel
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(CityPhotoViewModel::class)
-    internal abstract fun bindCityPhotoViewModel(cityPhotoViewModel: CityPhotoViewModel): ViewModel
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(SearchStartViewModel::class)
-    internal abstract fun bindSearchStartViewModel(searchStartViewModel: SearchStartViewModel): ViewModel
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(CityMapViewModel::class)
-    internal abstract fun bindCityMapViewModel(cityMapViewModel: CityMapViewModel): ViewModel
-
-    @Binds
-    internal abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
-}
