+++ app/src/main/java/com/xplorer/projectx/ui/MainActivity.kt
@@ -16,14 +16,16 @@
 package com.xplorer.projectx.ui
 
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.xplorer.projectx.R
 import com.xplorer.projectx.databinding.ActivityMainBinding
-import dagger.android.support.DaggerAppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
 import okhttp3.logging.HttpLoggingInterceptor
 import timber.log.Timber
 
-class MainActivity : DaggerAppCompatActivity() {
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
 
   private lateinit var binding: ActivityMainBinding
