+++ app/src/main/java/com/xplorer/projectx/ui/city/CityPhotoFragment.kt
@@ -22,8 +22,8 @@ import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.xplorer.projectx.R
@@ -32,17 +32,17 @@ import com.xplorer.projectx.model.CityModel
 import com.xplorer.projectx.model.unsplash.Photo
 import com.xplorer.projectx.ui.PhotoClickListener
 import com.xplorer.projectx.ui.adapter.CityPhotoPagedRecyclerAdapter
-import dagger.android.support.DaggerFragment
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
 /**
  * A simple [Fragment] subclass.
  */
-class CityPhotoFragment : DaggerFragment() {
+
+@AndroidEntryPoint
+class CityPhotoFragment : Fragment() {
   internal lateinit var binding: FragmentCityPhotoBinding
-  @Inject
-  lateinit var viewModelFactory: ViewModelProvider.Factory
-  private lateinit var viewModel: CityPhotoViewModel
+
+  private val viewModel: CityPhotoViewModel by viewModels()
   private lateinit var place: CityModel
   private lateinit var navController: NavController
 
@@ -53,8 +53,6 @@ class CityPhotoFragment : DaggerFragment() {
   ): View? {
     // Inflate the layout for this fragment
     binding = DataBindingUtil.inflate(inflater, R.layout.fragment_city_photo, container, false)
-    viewModel = ViewModelProvider(this, viewModelFactory)
-      .get(CityPhotoViewModel::class.java)
     displayPictures()
     (activity as AppCompatActivity).setSupportActionBar(binding.cityFragmentToolBar)
     (activity as AppCompatActivity).supportActionBar!!.setDisplayHomeAsUpEnabled(true)
