+++ app/src/main/java/com/xplorer/projectx/ui/search/SearchStartFragment.kt
@@ -22,6 +22,7 @@ import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
@@ -37,21 +38,22 @@ import com.xplorer.projectx.databinding.FragmentSearchStartBinding
 import com.xplorer.projectx.model.CityModel
 import com.xplorer.projectx.ui.adapter.recent.RecentCitiesAdapter
 import com.xplorer.projectx.utils.toCityModel
-import dagger.android.support.DaggerFragment
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
 /**
  * A simple [Fragment] subclass.
  */
-class SearchStartFragment : DaggerFragment() {
+
+
+@AndroidEntryPoint
+class SearchStartFragment : Fragment() {
 
   private lateinit var binding: FragmentSearchStartBinding
   private lateinit var placesClients: PlacesClient
 
   lateinit var navController: NavController
 
-  @Inject
-  lateinit var searchViewModel: SearchStartViewModel
+  private val searchViewModel: SearchStartViewModel by viewModels()
   private lateinit var recentCityAdapter: RecentCitiesAdapter
   private val recentCities = ArrayList<CityModel>()
