+++ app/src/main/java/com/xplorer/projectx/ui/search/SearchStartViewModel.kt
@@ -15,6 +15,7 @@
 */
 package com.xplorer.projectx.ui.search
 
+import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -38,7 +39,7 @@ import javax.inject.Inject
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-class SearchStartViewModel @Inject constructor(
+class SearchStartViewModel @ViewModelInject constructor(
   private val recentCitiesRepo: RecentCitiesRepo
 ) : ViewModel() {
