--- app/src/main/java/com/xplorer/projectx/view_model/ViewModelFactory.kt
@@ -1,49 +1,0 @@
-/**
- *  Designed and developed by ProjectX
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-package com.xplorer.projectx.view_model
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-import javax.inject.Singleton
-
-@Suppress("UNCHECKED_CAST")
-@Singleton
-class ViewModelFactory @Inject
-constructor(private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>) :
-  ViewModelProvider.Factory {
-
-  override fun <T : ViewModel> create(modelClass: Class<T>): T {
-    var creator: Provider<out ViewModel>? = creators[modelClass]
-    if (creator == null) {
-      for ((key, value) in creators) {
-        if (modelClass.isAssignableFrom(key)) {
-          creator = value
-          break
-        }
-      }
-    }
-    if (creator == null) {
-      throw IllegalArgumentException("unknown model class $modelClass")
-    }
-    try {
-      return creator.get() as T
-    } catch (e: Exception) {
-      throw RuntimeException(e)
-    }
-  }
-}
