+++ build.gradle
@@ -13,6 +13,7 @@ buildscript {
         classpath 'com.android.tools.build:gradle:' + versions.gradlePlugin
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:' + versions.kotlin
         classpath 'com.diffplug.spotless:spotless-plugin-gradle:' + versions.spotlessPlugin
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:' + versions.daggerHiltAndroidVersion
         classpath 'com.github.ben-manes:gradle-versions-plugin:' + versions.versionPlugin
         classpath 'com.google.gms:google-services:' + versions.google_services
         // NOTE: Do not place your application dependencies here; they belong
