+++ dependencies.gradle
@@ -10,7 +10,7 @@ ext.versions = [
         versionPlugin               : '0.27.0',
 
         // AndroidX
-        androidxCore                : '1.1.0',
+        androidxCore                : '1.2.0',
         androidxAnnotation          : '1.1.0',
         androidxRecyclerView        : '1.1.0',
         androidxMaterial            : '1.0.0',
@@ -23,6 +23,8 @@ ext.versions = [
 
         //dagger2
         dagger                      : '2.26',
+        daggerHiltAndroidVersion    : '2.28.3-alpha',
+        daggerHiltVersion           : '1.0.0-alpha02',
 
         //rx
         rxjava                      : '2.2.17',
